+++ sql/item_func.cc
@@ -2840,7 +2840,7 @@ longlong Item_func_locate::val_int()
   {
     start0= start= args[2]->val_int() - 1;
 
-    if ((start < 0) || (start > a->length()))
+    if ((start < 0) || ((size_t) start > a->length()))
       return 0;
 
     /* start is now sufficiently valid to pass to charpos function */
